requirements.txt
APScheduler==3.6.3
black==22.12.0
cachetools==4.2.2
certifi==2022.12.7
click==8.1.3
mypy-extensions==0.4.3
pathspec==0.10.3
platformdirs==2.6.2
python-dotenv==0.21.0
python-telegram-bot==13.15
pytz==2022.7
pytz-deprecation-shim==0.1.0.post0
six==1.16.0
tornado==6.1
tzdata==2022.7
tzlocal==4.2
